Add test to check InvalidWordStyleError raised
Inside the format_string function we were supposed to be raising
InvalidWordStyleError when the user passed in a wordstyle that isn't
supported.

I was calling the InvalidWordStyleError constructor but wasn't actually
raising the exception. This commit adds the test and fixes the
format_string function to actually raise the exception.