Simplify .coveragerc so Travis CI won't complain
Travis CI is saying my Python2.7 build is failing because it can't
parse the .coveragerc file provided. I've simplified the exclude_lines
entry to just be one line in hopes that will help it out.

I also added a leading comment in the file (because that's what the
documentation shows). And I added a branch checking option to increase
the level of detailed offered by coverage.