Remove print() when --initials and --count defined
I would like this tool to be useful in a `pipe` context. I.e. you
should be able to do:

    python namealizer.py > my_words.txt

With a reasonable expectation of not getting a bunch of debug messages.
To that end, it didn't make sense to warn the user that suplying both
--initials and --count would only result in --initials being used.

I now have changed the code to use the logging module, and if this
happens it is instead logged at the `info` level, which does not go to
STDIN by default.