Test getting word when starting_letter not in dict
This commit adds a test that verifies the function `get_random_word`
will raise a NoWordForLetter exception when given a `starting_letter`
that doesn't exist in the given dictionary.

In addition to the test, I implemented this functionality by adding a
try-except block which catches KeyErrors when trying to access the
dictionary.

Also had to define the NoWordForLetter exception.