Fix the majority of issues found by Code Climate
Code Climate identified 21 issues on its first build. I've installed a
local docker instance of the tool so I can use it to validate my builds
before pushing them.

There is still one unresolved issue of 'too high complexity in main'
but that is going to be a more invovled refactor so I want it to be on
it's own changeset so I can validate the changes separately.