Fix for first word in letter group imports as list
When working on the duplicate remover I noticed that a lot of extra
single-letter-words were being generated. After diving into the issue a
little bit with a smaller dictionary I realized that this was due to a
bug in the dictionary importer.

Given a list of words like:

    aardvark
    antietam
    antonym

These would be imported in as the following list:

`['a', 'a', 'r', 'd', 'v', 'a', 'r', 'k', 'antietam', 'antonym']`

The problem line was in `import_dictionary()`. In most cases when we
are adding a word to our internal dictionary we can append it to the
list that already exists for that letter, so for example:

    dictionary["a"].append(new_word)

However, when a word is added that doesn't already have a list to
append to, we need to create that list. That was being done like:

    dictionary[letter] = list(word)

Unfortunately that makes a new list out of the characters of the given
word. The correct way to just create a list with a single entry is:

    dictionary[letter] = [word]

I've also added two new checks in existing tests which will verify that
this issue is not reintroduced.

Fixes: Issue #17