Add namealizer.WordGenerator class
This commit is the first in a series of attempts to make this project
more useful as a Python module. It adds the `namealizer.WordGenerator`
class which is fairly simple and implements all of the same
functionality available from the command line. The object is created as
follows:

    wg = namealizer.WordGenerator("path_to_dictionary")

With other optional initializers of:

    wordstyle -> "lowercase", "uppercase", "capitalize", "mixedcase"
    separator -> Any ascii character to use as a word separator
    seed -> Number to seed the PRNG with, or None to auto-generate

You can then retrieve words in one of two ways:

    wg["abc"] -> returns three words where the words start with each
                 letter given. Raises an NoWordForLetter exception if a
                 word with the desired start letter doesn't exist.

    wg[3] -> returns three words where each word starts with a random
             letter

The `wordstyle` or `separator` can the be changed by accessing these
properties directly, such as:

    wg.separator = "-"
    wg.wordstyle = "uppercase"

Related to issue #6